+++ ft_memcmp.c
@@ -6,16 +6,19 @@
 /*   By: saichaou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 16:08:06 by saichaou          #+#    #+#             */
-/*   Updated: 2022/11/13 11:38:06 by saichaou         ###   ########.fr       */
+/*   Updated: 2022/11/15 14:52:03 by saichaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	int	i;
 
 	i = 0;
-	while (i < n && ((unsigned char *) s1)[i] == ((unsigned char *) s2)[i])
+	while (i < (int) n && ((unsigned char *) s1)[i]
+		== ((unsigned char *) s2)[i])
 		i++;
 	return (((unsigned char *)s1)[i] - ((unsigned char *)s2)[i]);
 }
