+++ ft_memcpy.c
@@ -6,21 +6,21 @@
 /*   By: saichaou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 13:53:29 by saichaou          #+#    #+#             */
-/*   Updated: 2022/11/12 16:11:06 by saichaou         ###   ########.fr       */
+/*   Updated: 2022/11/15 14:59:30 by saichaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "lib_ft.h"
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	int	i;
 
 	i = 0;
-	while (i < n)
+	while (i < (int) n)
 	{
 		((unsigned char *) dest)[i] = ((unsigned char *) src)[i];
-		i++;	
+		i++;
 	}
 	return (dest);
 }
